DASH-1326 Rotate apiUrl on error. Update tests.
