+++ lib/transactions/Request.js
@@ -320,6 +320,7 @@ class Request {
             const res = yield (0, utils_1.asyncWaterfall)({
                 asyncFunctions: this.config.baseUrls.map((apiUrl) => (signal) => this.executeCall({ baseUrl: apiUrl, endPoint: endpoint, body, fetchOptions, signal })),
                 requestTimeout,
+                baseUrls: this.config.baseUrls,
             });
             // TODO asyncWaterfall can throw errors and error interface can be different
             if (res === null || res === void 0 ? void 0 : res.isError) {
