+++ lib/utils/utils.js
@@ -21,7 +21,7 @@ const text_encoding_1 = require("text-encoding");
 const constants_1 = require("./constants");
 const DEFAULT_REQUEST_TIMEOUT = 60000;
 function asyncWaterfall(_a) {
-    return __awaiter(this, arguments, void 0, function* ({ asyncFunctions, requestTimeout = DEFAULT_REQUEST_TIMEOUT, }) {
+    return __awaiter(this, arguments, void 0, function* ({ asyncFunctions, requestTimeout = DEFAULT_REQUEST_TIMEOUT, baseUrls, }) {
         const timeoutIds = [];
         try {
             for (let i = 0; i < asyncFunctions.length; i++) {
@@ -47,11 +47,19 @@ function asyncWaterfall(_a) {
                 }
                 catch (error) {
                     clearTimeout(timeoutId);
-                    if ((error === null || error === void 0 ? void 0 : error.code) && [constants_1.API_ERROR_CODES.NOT_FOUND, constants_1.API_ERROR_CODES.BAD_REQUEST].includes(error.code)
-                        || (error === null || error === void 0 ? void 0 : error.errorCode) && [constants_1.API_ERROR_CODES.NOT_FOUND, constants_1.API_ERROR_CODES.BAD_REQUEST].includes(error.errorCode)) {
+                    const errorCode = (error === null || error === void 0 ? void 0 : error.code) || (error === null || error === void 0 ? void 0 : error.errorCode);
+                    if (errorCode && [constants_1.API_ERROR_CODES.NOT_FOUND, constants_1.API_ERROR_CODES.BAD_REQUEST].includes(errorCode)) {
                         throw error;
                     }
-                    if (i === asyncFunctions.length - 1) {
+                    // If this isn't the last function and we have baseUrls, reorder them
+                    if (baseUrls && i !== asyncFunctions.length - 1) {
+                        // Move failed URL to the end and next call will use the next url from the list not thefailed one
+                        // Works only for current instance of the sdk
+                        const failedUrl = baseUrls[i];
+                        baseUrls.splice(i, 1);
+                        baseUrls.push(failedUrl);
+                    }
+                    else if (i === asyncFunctions.length - 1) {
                         throw error;
                     }
                 }
