+++ tests/testnet.spec.ts
@@ -3,8 +3,9 @@ import chai from 'chai';
 import * as dotenv from 'dotenv'
 import 'mocha'
 import nodeFetch from 'node-fetch'
-// Add import for 400 error test
-import { test400Error } from './common/400-error.spec'
+
+import { wrongRequestsTests } from './common/wrong-requests.spec'
+
 import {
     Account,
     Action,
@@ -16,13 +17,11 @@ import {
     TransactionResponse,
 } from '../src/FIOSDK'
 
-import type { FioError as FioErrorType } from '../src/entities/types/FioError.d.ts';
-
 dotenv.config({path: ['.env.test', '.env']})
 
 const { expect } = chai;
 
-type ErrorType = Error & FioErrorType;
+import type { ErrorType } from './utils';
 
 const fetchJson = async (uri: string, opts = {}) => {
     return nodeFetch(uri, opts)
@@ -62,8 +61,6 @@ const ethTokenCode = 'ETH'
 const ethChainCode = 'ETH'
 const defaultBundledSets = 1
 const defaultFee = 1500 * FIOSDK.SUFUnit
-const wrongBaseUrl = 'https://wrong-url-test.test.com/'
-const wrongBaseUrl2 = 'https://wrong-url-test-2.com/'
 
 let fioSdk: FIOSDK = new FIOSDK({ privateKey, publicKey, apiUrls: baseUrls, fetchJson })
 let fioSdk2: FIOSDK = new FIOSDK({ privateKey: privateKey2, publicKey: publicKey2, apiUrls: baseUrls, fetchJson })
@@ -183,130 +180,7 @@ describe('Raw Abi missing', () => {
     })
 })
 
-test400Error({ fioSdk, baseUrls })
-
-describe('Testing request timeout on wrong url', () => {
-    it(`Get Fio Balance with wrong base url`, async () => {
-        try {
-            fioSdkWithWrongBaseUrl.setApiUrls([wrongBaseUrl])
-            await fioSdkWithWrongBaseUrl.genericAction('getFioBalance', {})
-        } catch (e) {
-            if (e instanceof Error) {
-                expect(e.message).to.match(/request_timeout|ENOTFOUND/)
-            }
-        }
-    })
-
-    it(`Get Fio Balance with 2 wrong base urls`, async () => {
-        try {
-            fioSdkWithWrongBaseUrl.setApiUrls([wrongBaseUrl, wrongBaseUrl2])
-            await fioSdkWithWrongBaseUrl.genericAction('getFioBalance', {})
-        } catch (e) {
-            if (e instanceof Error) {
-                expect(e.message).to.match(/request_timeout|ENOTFOUND/)
-            }
-        }
-    })
-
-    it(`Get Fio Balance with one wrong and correct base urls`, async () => {
-        fioSdkWithWrongBaseUrl.setApiUrls([wrongBaseUrl, ...baseUrls])
-
-        const result = await fioSdkWithWrongBaseUrl.genericAction(
-            'getFioBalance',
-            {},
-        )
-        expect(result).to.have.all.keys(
-            'balance',
-            'available',
-            'staked',
-            'srps',
-            'roe',
-        )
-        expect(result.balance).to.be.a('number')
-        expect(result.available).to.be.a('number')
-        expect(result.staked).to.be.a('number')
-        expect(result.srps).to.be.a('number')
-        expect(result.roe).to.be.a('string')
-    })
-
-    it(`Get Fio Balance with one correct and wrong base urls`, async () => {
-        fioSdkWithWrongBaseUrl.setApiUrls([...baseUrls, wrongBaseUrl])
-
-        const result = await fioSdkWithWrongBaseUrl.genericAction(
-            'getFioBalance',
-            {},
-        )
-        expect(result).to.have.all.keys(
-            'balance',
-            'available',
-            'staked',
-            'srps',
-            'roe',
-        )
-        expect(result.balance).to.be.a('number')
-        expect(result.available).to.be.a('number')
-        expect(result.staked).to.be.a('number')
-        expect(result.srps).to.be.a('number')
-        expect(result.roe).to.be.a('string')
-    })
-
-    it(`Get Fio Balance with one wrong, correct and wrong base urls`, async () => {
-        fioSdkWithWrongBaseUrl.setApiUrls([
-            wrongBaseUrl,
-            baseUrls[0],
-            wrongBaseUrl2,
-        ])
-
-        const result = await fioSdkWithWrongBaseUrl.genericAction(
-            'getFioBalance',
-            {},
-        )
-        expect(result).to.have.all.keys(
-            'balance',
-            'available',
-            'staked',
-            'srps',
-            'roe',
-        )
-        expect(result.balance).to.be.a('number')
-        expect(result.available).to.be.a('number')
-        expect(result.staked).to.be.a('number')
-        expect(result.srps).to.be.a('number')
-        expect(result.roe).to.be.a('string')
-    })
-
-    it(`Make removePublicAddresses request with wrong parameter and correct base url`, async () => {
-        fioSdkWithWrongBaseUrl.setApiUrls(baseUrls)
-        try {
-            await fioSdk.genericAction('removePublicAddresses', {
-                fioAddress: '',
-                maxFee: 600000000,
-                publicAddresses: [
-                    {
-                        chain_code: 'BCH',
-                        public_address:
-                            'bitcoincash:qzf8zha74ahdh9j0xnwlffdn0zuyaslx3c90q7n9g9',
-                        token_code: 'BCH',
-                    },
-                    {
-                        chain_code: 'DASH',
-                        public_address: 'XyCyPKzTWvW2XdcYjPaPXGQDCGk946ywEv',
-                        token_code: 'DASH',
-                    },
-                ],
-            })
-        } catch (err) {
-            const error = err as ErrorType;
-
-            expect(error.message).to.equal('Validation error');
-            expect(error.list[0].message).to.equal('fioAddress must have a length between 3 and 64.')
-        }
-    })
-
-    it(`Return back correct baseUrls`, () => {
-        fioSdkWithWrongBaseUrl.setApiUrls(baseUrls)
-    })
-})
+wrongRequestsTests({ fioSdk, baseUrls, fioSdkWithWrongBaseUrl })
 
 describe('Testing generic actions', () => {
     const newFioDomain = generateTestingFioDomain()
