+++ tests/utils.ts
@@ -1,0 +1,3 @@
+import type { FioError as FioErrorType } from '../src/entities/types/FioError.d.ts';
+
+export type ErrorType = Error & FioErrorType;
